Restructure project to support collection review
